Add a test for checking if prometheus operator schema is correct
