+++ .circleci/config.yml
@@ -6,11 +6,11 @@ defaults: &defaults
     GRUNTWORK_INSTALLER_VERSION: v0.0.21
     TERRATEST_LOG_PARSER_VERSION: v0.13.13
     HELM_VERSION: v2.12.2
-    MODULE_CI_VERSION: v0.14.1
+    MODULE_CI_VERSION: v0.16.0
     TERRAFORM_VERSION: NONE
     TERRAGRUNT_VERSION: NONE
     PACKER_VERSION: NONE
-    GOLANG_VERSION: 1.11.2
+    GOLANG_VERSION: 1.13
     KUBECONFIG: /home/circleci/.kube/config
 
 
@@ -32,13 +32,10 @@ install_gruntwork_utils: &install_gruntwork_utils
     gruntwork-install --module-name "kubernetes-circleci-helpers" --repo "https://github.com/gruntwork-io/module-ci" --tag "${MODULE_CI_VERSION}"
     gruntwork-install --binary-name "terratest_log_parser" --repo "https://github.com/gruntwork-io/terratest" --tag "${TERRATEST_LOG_PARSER_VERSION}"
     configure-environment-for-gruntwork-module \
-      --circle-ci-2-machine-executor \
       --terraform-version ${TERRAFORM_VERSION} \
       --terragrunt-version ${TERRAGRUNT_VERSION} \
       --packer-version ${PACKER_VERSION} \
-      --use-go-dep \
       --go-version ${GOLANG_VERSION} \
-      --go-src-path test
 
 
 version: 2
@@ -47,9 +44,6 @@ jobs:
     <<: *defaults
     steps:
       - checkout
-      - restore_cache:
-          keys:
-          - dep-{{ checksum "test/Gopkg.lock" }}
 
       # Install gruntwork utilities
       - run:
@@ -57,11 +51,6 @@ jobs:
 
       - run:
           <<: *install_helm_client
-
-      - save_cache:
-          key: dep-{{ checksum "test/Gopkg.lock" }}
-          paths:
-          - ./test/vendor
 
       # Fail the build if the pre-commit hooks don't pass. Note: if you run pre-commit install locally, these hooks will
       # execute automatically every time before you commit, ensuring the build never fails at this step!
@@ -97,7 +86,9 @@ jobs:
           name: run tests
           command: |
             mkdir -p /tmp/logs
-            run-go-tests --packages "-tags all ./test" --timeout 60m | tee /tmp/logs/all.log
+            cd test
+            go mod tidy
+            run-go-tests --packages "-tags all ." --timeout 60m | tee /tmp/logs/all.log
           no_output_timeout: 3600s
 
       - run:
