+++ test/go.mod
@@ -1,0 +1,13 @@
+module github.com/gruntwork-io/helm-kubernetes-services/test
+
+go 1.13
+
+require (
+	github.com/GoogleCloudPlatform/gke-managed-certs v0.3.4
+	github.com/coreos/prometheus-operator v0.33.1-0.20191023063855-29199d546e0b
+	github.com/ghodss/yaml v1.0.0
+	github.com/gruntwork-io/terratest v0.22.2
+	github.com/stretchr/testify v1.4.0
+	k8s.io/api v0.0.0-20190918155943-95b840bb6a1f
+	k8s.io/apimachinery v0.0.0-20190913080033-27d36303b655
+)
