+++ test/k8s_service_example_test_helpers.go
@@ -77,6 +77,7 @@ func verifySinglePodAvailable(
 	http_helper.HttpGetWithRetryWithCustomValidation(
 		t,
 		fmt.Sprintf("http://%s", tunnel.Endpoint()),
+		nil,
 		WaitTimerRetries,
 		WaitTimerSleep,
 		validationFunction,
@@ -108,6 +109,7 @@ func verifyServiceAvailable(
 	http_helper.HttpGetWithRetryWithCustomValidation(
 		t,
 		fmt.Sprintf("http://%s", serviceEndpoint),
+		nil,
 		WaitTimerRetries,
 		WaitTimerSleep,
 		validationFunction,
