+++ test/k8s_service_template_test.go
@@ -62,28 +62,6 @@ func TestK8SServiceServiceEnabledFalseDoesNotCreateService(t *testing.T) {
 	assert.Equal(t, len(rendered), 0)
 }
 
-// Test that setting serviceMonitor.enabled = false will cause the helm template to not render the Service Monitor resource
-func TestK8SServiceServiceMonitorEnabledFalseDoesNotCreateServiceMonitor(t *testing.T) {
-	t.Parallel()
-
-	helmChartPath, err := filepath.Abs(filepath.Join("..", "charts", "k8s-service"))
-	require.NoError(t, err)
-
-	// We make sure to pass in the linter_values.yaml values file, which we assume has all the required values defined.
-	// We then use SetValues to override all the defaults.
-	options := &helm.Options{
-		ValuesFiles: []string{filepath.Join("..", "charts", "k8s-service", "linter_values.yaml")},
-		SetValues:   map[string]string{"serviceMonitor.enabled": "false"},
-	}
-	out := helm.RenderTemplate(t, options, helmChartPath, []string{"templates/servicemonitor.yaml"})
-
-	// We take the output and render it to a map to validate it is an empty yaml
-	rendered := map[string]interface{}{}
-	err = yaml.Unmarshal([]byte(out), &rendered)
-	assert.NoError(t, err)
-	assert.Equal(t, len(rendered), 0)
-}
-
 // Test each of the required values. Here, we take advantage of the fact that linter_values.yaml is supposed to define
 // all the required values, so we check the template rendering by nulling out each field.
 func TestK8SServiceRequiredValuesAreRequired(t *testing.T) {
